Move pseudo-annotation completions to pseudo package
